investbud-core/gcn/pyproject.toml
[project]
name = "investbud-gcn"
version = "0.1.0"
description = "InvestBud AI Graph Confirming Node - Independent verifier for DKG oracle snapshots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dkg>=8.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
full = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "fredapi>=0.5.0",
    "yfinance>=0.2.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
gcn = "investbud_gcn.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
